:truck: move AppComponent to pages/top
